refactor(aicr): move public facade from module root to pkg/aicr/
The top-level facade (Client, NewClient, RecipeResult, request/response
types, ValidateOption, Snapshot/AgentConfig/PhaseResult aliases, source
constructors) lived in the module root from #1072 onward. That layout
puts library code, binary commands (cmd/aicr, cmd/aicrd), and tooling
side-by-side at the repository root and reads as ambiguous for a
multi-binary, multi-library project — most Go repos in this shape keep
the library packages under pkg/ and reserve the module root for go.mod,
top-level docs, and Makefile only.

This commit moves the 5 root facade files into pkg/aicr/ verbatim
(git mv preserves history):

- aicr.go                  → pkg/aicr/aicr.go
- aicr_internal_test.go    → pkg/aicr/aicr_internal_test.go
- aicr_test.go             → pkg/aicr/aicr_test.go
- options.go               → pkg/aicr/options.go
- types.go                 → pkg/aicr/types.go

The package name (aicr) is unchanged; only the IMPORT PATH changes.

**BREAKING for external Go importers.** Consumers must update:

  - import "github.com/NVIDIA/aicr"
  + import aicr "github.com/NVIDIA/aicr/pkg/aicr"

In-tree call sites are unaffected: no in-tree code imports the root
facade today (the CLI + REST server still construct recipe.Builder /
bundler.Bundler directly; that migration is #1108).

Docs updated: docs/integrator/go-library.md and
docs/integrator/public-api.md reflect the new import path. The Go
MODULE name (github.com/NVIDIA/aicr from go.mod) is unchanged; SBOM
entries and GitHub repo URLs continue to use it.

Related: #1076 (facade-hardening epic; unblocks cleaner #1108 and #1111
rebases by establishing the canonical location before they land).
